Add a 'Component' trait for vector components
